backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# AI & LLM
anthropic==0.41.0
openai==1.58.1

# Document Processing
python-docx==1.1.2
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# RAG
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-openai==0.2.14
chromadb==0.5.23
tiktoken==0.8.0

# Agents
langgraph==0.2.53

# Utilities
python-dotenv==1.0.1